* deleted forms from admin.py * added forms to forms.py #28
